refactoring: separated auth
